Monte Carlo Work I should really work on Multiplayer

+++ src/sample/AI/Move.java
@@ -6,10 +6,10 @@ import java.awt.*;
  * For use by the Minimax algorithm bundles together a move and that moves value.
  */
 class Move {
-    public int value;
+    public double value;
     public final Point space;
 
-    public Move(Point space, int value) {
+    public Move(Point space, double value) {
         this.space = space;
         this.value = value;
     }
